requirements.txt
#atdlib==0.8a8
bitcoin==1.1.42
configobj==5.0.6
Django==1.10.4
ecdsa==0.13
functools32==3.2.3.post2
hidapi==0.7.99.post20
jsonschema==2.5.1
mnemonic==0.15
pbkdf2==1.3
protobuf==3.1.0.post1
pycrypto==2.6.1
pypreprocessor==0.6.0
ruamel.ordereddict==0.4.9
ruamel.yaml==0.13.7
selenium==3.0.2
simplejson==3.10.0
six==1.10.0
trezor==0.7.8
typing==3.5.2.2
web.py==0.38
